Makefile: Trap errors with -Wall, -Werror compiler flags for Linux also (#13528)
* Makefile: Trap errors with -Wall, -Werror compiler flags for Linux also

Squashed:
	pnso_chain: Fixup warning/error in Linux build
	PenCAKE: Exclude emitter.c to avoid warning/error at compilation time


* PenCAKE: fix stack size compile warning

+++ storage/offload/src/drv/pnso_chain.c
@@ -814,8 +814,9 @@ hw_ring_evaluate_take(struct service_info *svc_info)
 		return;
 	}
 
-	ring_tracker->rt_max_dflt_takes = max(ring_tracker->rt_max_dflt_takes,
-				svc_info->si_seq_info.sqi_hw_dflt_takes);
+	ring_tracker->rt_max_dflt_takes =
+		max((uint8_t) ring_tracker->rt_max_dflt_takes,
+			(uint8_t) svc_info->si_seq_info.sqi_hw_dflt_takes);
 
 	svc_info->si_seq_info.sqi_hw_dflt_takes = 0;
 }
